+++ platform.h
@@ -8,7 +8,7 @@ class Platform : public sf::RectangleShape
 public:
     Platform();
     Platform(int x, int y, sf::Vector2f size, sf::Vector2f texture_size, std::string texture_name);
-    void repeate(int x, int y = 1);
+    void repeate(float x, float y = 1);
     static void platform_maker(sf::Vector2f quantity, sf::Vector2f size, sf::Vector2f screen,
     sf::Vector2f texture_size, std::string texture_name, sf::Vector2f texture_multi, std::vector<std::unique_ptr<sf::Drawable>> &vector);
     sf::Vector2f get_size();
